requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
crispy-bootstrap5==2025.6
cryptography==46.0.5
Django==5.2.11
django-cors-headers==4.9.0
django-crispy-forms==2.5
django-environ==0.13.0
djangorestframework==3.16.1
Flask==3.1.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.190.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.78.1
grpcio-status==1.71.2
gunicorn==25.1.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
MarkupSafe==3.0.2
mysql-connector-python==9.3.0
nltk==3.9.1
packaging==26.0
pillow==12.1.1
proto-plus==1.27.1
protobuf==5.29.6
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyjokes==0.8.3
pyparsing==3.3.2
PyPDF2==3.0.1
python-dotenv==1.2.1
regex==2024.11.6
requests==2.32.5
sqlparse==0.5.5
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
Werkzeug==3.1.3
whitenoise==6.12.0
youtube-search-python==1.6.6
